Reformatted source files.
Modified project description.

Updated README.

Cleaned up project files.

Formatted pom.xml.
